fix: move syncDataType and IntegrationCategoryType enum to noco-sdk
